requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.2.1
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
editor==1.6.6
environs==14.1.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
marshmallow==4.0.0
moviepy==2.1.2
numpy==2.2.5
openapi==3.1.1
packaging==25.0
pillow==11.2.1
proglog==0.1.11
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
runs==1.2.2
self==2020.12.3
setuptools==79.0.1
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
standard-distutils==3.11.9
stripe==12.0.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
xmod==1.8.1